# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CanteenFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  student: "student",
  canteen: "canteen",
} as const;

export interface UserProfile {
  id: number;
  clerkId: string;
  name: string;
  email: string;
  role: UserProfileRole;
  createdAt: string;
}

export type UserUpdateRole =
  (typeof UserUpdateRole)[keyof typeof UserUpdateRole];

export const UserUpdateRole = {
  student: "student",
  canteen: "canteen",
} as const;

export interface UserUpdate {
  name?: string;
  role?: UserUpdateRole;
}

export interface MenuItem {
  id: number;
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  prepTimeMinutes: number;
  isAvailable: boolean;
  isVeg: boolean;
  /** @nullable */
  calories: number | null;
  createdAt: string;
}

export interface MenuItemInput {
  name: string;
  description: string;
  price: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  prepTimeMinutes: number;
  isAvailable: boolean;
  isVeg: boolean;
  /** @nullable */
  calories?: number | null;
}

export interface MenuItemUpdate {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  prepTimeMinutes?: number;
  isAvailable?: boolean;
  isVeg?: boolean;
  /** @nullable */
  calories?: number | null;
}

export interface OrderItem {
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  price: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  confirmed: "confirmed",
  cooking: "cooking",
  packaging: "packaging",
  ready: "ready",
  collected: "collected",
  cancelled: "cancelled",
} as const;

export type OrderType = (typeof OrderType)[keyof typeof OrderType];

export const OrderType = {
  dine_in: "dine_in",
  parcel: "parcel",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  upi: "upi",
  card: "card",
  cod: "cod",
  razorpay: "razorpay",
} as const;

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export interface Order {
  id: number;
  userId: number;
  userName: string;
  items: OrderItem[];
  totalAmount: number;
  status: OrderStatus;
  type: OrderType;
  /** @nullable */
  groupSize: number | null;
  /** @nullable */
  tableNumber: string | null;
  paymentMethod: OrderPaymentMethod;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  estimatedReadyAt: string | null;
  createdAt: string;
  updatedAt: string;
}

export type OrderInputItemsItem = {
  menuItemId: number;
  quantity: number;
};

export type OrderInputType =
  (typeof OrderInputType)[keyof typeof OrderInputType];

export const OrderInputType = {
  dine_in: "dine_in",
  parcel: "parcel",
} as const;

export type OrderInputPaymentMethod =
  (typeof OrderInputPaymentMethod)[keyof typeof OrderInputPaymentMethod];

export const OrderInputPaymentMethod = {
  upi: "upi",
  card: "card",
  cod: "cod",
  razorpay: "razorpay",
} as const;

export interface OrderInput {
  items: OrderInputItemsItem[];
  type: OrderInputType;
  /** @nullable */
  groupSize?: number | null;
  paymentMethod: OrderInputPaymentMethod;
}

export type OrderStatusUpdateStatus =
  (typeof OrderStatusUpdateStatus)[keyof typeof OrderStatusUpdateStatus];

export const OrderStatusUpdateStatus = {
  confirmed: "confirmed",
  cooking: "cooking",
  packaging: "packaging",
  ready: "ready",
  collected: "collected",
  cancelled: "cancelled",
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export type OrderSummaryByStatus = {
  confirmed: number;
  cooking: number;
  packaging: number;
  ready: number;
  collected: number;
};

export interface OrderSummary {
  totalToday: number;
  activeNow: number;
  avgPrepTime: number;
  revenue: number;
  byStatus: OrderSummaryByStatus;
}

export type SeatStatus = (typeof SeatStatus)[keyof typeof SeatStatus];

export const SeatStatus = {
  available: "available",
  occupied: "occupied",
  reserved: "reserved",
} as const;

export interface Seat {
  id: number;
  seatNumber: string;
  tableNumber: string;
  zone: string;
  row: number;
  col: number;
  status: SeatStatus;
  /** @nullable */
  reservationId: number | null;
  /** @nullable */
  reservedUntil: string | null;
}

export interface Reservation {
  id: number;
  userId: number;
  seatIds: number[];
  tableNumber: string;
  seatNumbers: string[];
  groupSize: number;
  expiresAt: string;
  createdAt: string;
}

export type ReservationOrNull = Reservation | null;

export interface ReserveSeatsInput {
  groupSize: number;
  preferredZone?: string;
}

export type OccupancyStatsZonesItem = {
  zone: string;
  available: number;
  occupied: number;
  reserved: number;
};

export interface OccupancyStats {
  totalSeats: number;
  available: number;
  occupied: number;
  reserved: number;
  occupancyPercent: number;
  zones: OccupancyStatsZonesItem[];
}

export interface PaymentOrderInput {
  amount: number;
  currency: string;
  /** @nullable */
  orderId?: number | null;
}

export interface PaymentOrder {
  razorpayOrderId: string;
  amount: number;
  currency: string;
  keyId: string;
}

export interface PaymentVerifyInput {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
  /** @nullable */
  orderId?: number | null;
}

export interface PaymentVerifyResult {
  success: boolean;
  message: string;
}

export type AiRecommendationsItemsItem = {
  menuItemId: number;
  name: string;
  reason: string;
  prepTimeMinutes: number;
  price: number;
};

export interface AiRecommendations {
  message: string;
  items: AiRecommendationsItemsItem[];
}

export type CrowdPredictionCurrentLevel =
  (typeof CrowdPredictionCurrentLevel)[keyof typeof CrowdPredictionCurrentLevel];

export const CrowdPredictionCurrentLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  very_high: "very_high",
} as const;

export type CrowdPredictionPredictionsItemLevel =
  (typeof CrowdPredictionPredictionsItemLevel)[keyof typeof CrowdPredictionPredictionsItemLevel];

export const CrowdPredictionPredictionsItemLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  very_high: "very_high",
} as const;

export type CrowdPredictionPredictionsItem = {
  hour: number;
  minute: number;
  level: CrowdPredictionPredictionsItemLevel;
  occupancyPercent: number;
};

export interface CrowdPrediction {
  currentLevel: CrowdPredictionCurrentLevel;
  predictions: CrowdPredictionPredictionsItem[];
  recommendation: string;
}

export interface BestOrderTime {
  currentWaitMinutes: number;
  bestTimeMinutes: number;
  suggestion: string;
  breakTimeOptimized: boolean;
}

export type CanteenDashboardTopItemsItem = {
  name: string;
  count: number;
};

export interface CanteenDashboard {
  ordersToday: number;
  revenueToday: number;
  activeOrders: number;
  occupiedSeats: number;
  totalSeats: number;
  avgPrepTime: number;
  topItems: CanteenDashboardTopItemsItem[];
  recentOrders: Order[];
}

export type ListMenuItemsParams = {
  category?: string;
  available?: string;
};

export type ListOrdersParams = {
  status?: string;
  type?: string;
};