# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CanteenFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu:
    get:
      operationId: listMenuItems
      tags: [menu]
      summary: List menu items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current user (or all for canteen)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (canteen only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── SEATS ───────────────────────────────────────────────────────────────────
...
  /payments/create-order:
    post:
      operationId: createPaymentOrder
      tags: [payments]
      summary: Create a mock payment order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentOrderInput"
      responses:
        "201":
          description: Payment order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentOrder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/verify:
    post:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify payment signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentVerifyInput"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResult"
        "400":
          description: Invalid signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── AI AGENTS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        prepTimeMinutes:
          type: integer
        isAvailable:
          type: boolean
        isVeg:
          type: boolean
        calories:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - price
        - category
        - imageUrl
        - prepTimeMinutes
        - isAvailable
        - isVeg
        - calories
        - createdAt
    MenuItemInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        prepTimeMinutes:
          type: integer
        isAvailable:
          type: boolean
        isVeg:
          type: boolean
        calories:
          type: ["integer", "null"]
      required:
        - name
        - description
        - price
        - category
        - prepTimeMinutes
        - isAvailable
        - isVeg
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        prepTimeMinutes:
          type: integer
        isAvailable:
          type: boolean
        isVeg:
          type: boolean
        calories:
          type: ["integer", "null"]
    OrderItem:
      type: object
      properties:
        menuItemId:
          type: integer
        menuItemName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - menuItemId
        - menuItemName
        - quantity
        - price
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalAmount:
          type: number
        status:
          type: string
          enum: [confirmed, cooking, packaging, ready, collected, cancelled]
        type:
          type: string
          enum: [dine_in, parcel]
        groupSize:
          type: ["integer", "null"]
        tableNumber:
          type: ["string", "null"]
        paymentMethod:
          type: string
          enum: [upi, card, cod, razorpay]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        estimatedReadyAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userName
        - items
        - totalAmount
        - status
        - type
        - groupSize
        - tableNumber
        - paymentMethod
        - paymentStatus
        - estimatedReadyAt
        - createdAt
        - updatedAt
    OrderInput:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              menuItemId:
                type: integer
              quantity:
                type: integer
            required:
              - menuItemId
              - quantity
        type:
          type: string
          enum: [dine_in, parcel]
        groupSize:
          type: ["integer", "null"]
        paymentMethod:
          type: string
          enum: [upi, card, cod, razorpay]
      required:
        - items
        - type
        - paymentMethod
    OrderStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, cooking, packaging, ready, collected, cancelled]
      required:
        - status
...
    PaymentOrderInput:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
        orderId:
          type: ["integer", "null"]
      required:
        - amount
        - currency
    PaymentOrder:
      type: object
      properties:
        razorpayOrderId:
          type: string
        amount:
          type: number
        currency:
          type: string
        keyId:
          type: string
      required:
        - razorpayOrderId
        - amount
        - currency
        - keyId
    PaymentVerifyInput:
      type: object
      properties:
        razorpayOrderId:
          type: string
        razorpayPaymentId:
          type: string
        razorpaySignature:
          type: string
        orderId:
          type: ["integer", "null"]
      required:
        - razorpayOrderId
        - razorpayPaymentId
        - razorpaySignature
    PaymentVerifyResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...