# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CanteenFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderPaymentMethod } from "./orderPaymentMethod";
import type { OrderPaymentStatus } from "./orderPaymentStatus";
import type { OrderStatus } from "./orderStatus";
import type { OrderType } from "./orderType";

export interface Order {
  id: number;
  userId: number;
  userName: string;
  items: OrderItem[];
  totalAmount: number;
  status: OrderStatus;
  type: OrderType;
  /** @nullable */
  groupSize: number | null;
  /** @nullable */
  tableNumber: string | null;
  paymentMethod: OrderPaymentMethod;
  paymentStatus: OrderPaymentStatus;
  /** @nullable */
  estimatedReadyAt: Date | null;
  createdAt: Date;
  updatedAt: Date;
}