# lib/api-zod/src/generated/types/paymentOrderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CanteenFlow AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentOrderInput {
  amount: number;
  currency: string;
  /** @nullable */
  orderId?: number | null;
}